Update expected queries for changes in Django's ORM.
Django 4+ has made some under-the-hood changes to related fields in the
ORM, which broke some of our expected queries. This change fixes them up
so tests pass.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/13484/

+++ reviewboard/reviews/models/review_request.py
@@ -1175,7 +1175,7 @@ class ReviewRequest(BaseReviewRequestDetails):
         if update_counts or self.id is None:
             self._update_counts(old_submitter)
 
-        if self.status != self.PENDING_REVIEW:
+        if self.pk and self.status != self.PENDING_REVIEW:
             # If this is not a pending review request now, delete any
             # and all ReviewRequestVisit objects.
             self.visits.all().delete()
